% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_jssp_bibliography.R
\docType{data}
\name{jssp.bibliography}
\alias{jssp.bibliography}
\title{The Literature References from which we obtained the Results for the Job Shop Scheduling Problem}
\format{A data frame with the bibliographic references
\describe{
  \item{ref.id}{the reference id}
  \item{ref.type}{the bibtex-based type}
  \item{ref.year}{the year when the publication was created}
  \item{ref.as.bibtex}{the bibtex representation of the reference}
  \item{ref.as.text}{the textual representation of the reference}
}}
\usage{
data(jssp.bibliography)
}
\description{
This data frame holds the bibliographic information from which we obtained
the results for the Job Shop Scheduling Problem.
}
\examples{
data(jssp.bibliography)
print(jssp.bibliography)
}
\keyword{JSSP,}
\keyword{Job}
\keyword{Scheduling,}
\keyword{Shop}
\keyword{bibliography}
