% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_gantt.R
\name{jssp.evaluate.gantt}
\alias{jssp.evaluate.gantt}
\title{Check and Evaluate a Gantt Chart, i.e., A Candidate Solution for a
  Given JSSP Instance}
\usage{
jssp.evaluate.gantt(gantt, inst.id,
  get.inst.data = jssp.get.instance.data)
}
\arguments{
\item{gantt}{the gantt chart}

\item{inst.id}{the instance id}

\item{get.inst.data}{a function obtaining the instance data for a given
instance id, by default \link{jssp.get.instance.data}}
}
\value{
the evaluation result
}
\description{
This function accepts an instance id and a Gantt chart and
  verifies whether the Gantt chart is a proper solution and evaluates its
  makespan.
}
