% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_instance_data.R
\name{jssp.get.instance.data}
\alias{jssp.get.instance.data}
\title{Get the Data for a Given Instance}
\usage{
jssp.get.instance.data(inst.id)
}
\arguments{
\item{inst.id}{the id/name of the JSSP instance}
}
\value{
the instance data entry, formatted according to
  \link{jssp.instance.data}
}
\description{
For an instance name, get the corresponding data
}
