% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_gantt.R
\name{jssp.ob.to.gantt}
\alias{jssp.ob.to.gantt}
\title{Transform a Solution Represented in Form \code{OB} to a Gantt Chart}
\usage{
jssp.ob.to.gantt(data.ob, inst.id, min.job.id = 1L,
  get.inst.data = jssp.get.instance.data)
}
\arguments{
\item{data.ob}{the operation-based representation of the solution}

\item{inst.id}{the instance id}

\item{min.job.id}{the integer minimum job id to be used in the output (for
the input, it is automatically detected). By default, this is \code{1L},
but sometimes you may want to use \code{0L}.}

\item{get.inst.data}{a function obtaining the instance data for a given
instance id, by default \link{jssp.get.instance.data}}
}
\value{
the canonicalized and evaluated Gantt chart, see
  \link{jssp.evaluate.gantt}
}
\description{
A solution in the \code{OB} (operation-based) representation is
  given. each job is represented by \\code{m} genes with the same value and
  the chromosome is processed from front to end by assigning jobs to machines
  at the earliest starting times, following their occurence order. This
  solution is then transformed to a Gantt chart à la
  \link{jssp.evaluate.gantt}.
}
\seealso{
jssp.evaluate.gantt
}
