% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_gantt.R
\name{jssp.oo.to.gantt}
\alias{jssp.oo.to.gantt}
\title{Transform a Solution Represented in the van-Hoorn Version of \code{OO}
  to a Gantt Chart}
\usage{
jssp.oo.to.gantt(data.oo, inst.id, min.job.id = 1L,
  get.inst.data = jssp.get.instance.data)
}
\arguments{
\item{data.oo}{the overall-order based representation of the solution}

\item{inst.id}{the instance id}

\item{min.job.id}{the integer minimum job id to be used in the output (for
the input, it is automatically detected). By default, this is \code{1L},
but sometimes you may want to use \code{0L}.}

\item{get.inst.data}{a function obtaining the instance data for a given
instance id, by default \link{jssp.get.instance.data}}
}
\value{
the Gantt chart
}
\description{
A solution in the van-Hoorn version of \code{OO} is translated
  to a Gantt Chart. Here, the operations are numbered as follows: The first
  \code{n} operations refer to the first operation of each job (according to
  order of the jobs), operations \code{n+1,...,2n} regard the second
  operation of the \code{n} jobs, and so on. So operation \code{i} is the
  \code{k}'th operation of job \code{j}, where \code{k = ceil(i/n)} and
  \code{j = i mod n}.
}
