% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_gantt.R
\name{jssp.pl.to.gantt}
\alias{jssp.pl.to.gantt}
\title{Transform a Solution Represented in the \code{PL} Representation to a
  Gantt Chart}
\usage{
jssp.pl.to.gantt(data.pl, inst.id, min.job.id = 1L,
  get.inst.data = jssp.get.instance.data)
}
\arguments{
\item{data.pl}{the pl representation}

\item{inst.id}{the instance id}

\item{min.job.id}{the integer minimum job id to be used in the output (for
the input, it is automatically detected). By default, this is \code{1L},
but sometimes you may want to use \code{0L}.}

\item{get.inst.data}{a function obtaining the instance data for a given
instance id, by default \link{jssp.get.instance.data}}
}
\value{
the canonicalized and evaluated Gantt chart, see
  \link{jssp.evaluate.gantt}
}
\description{
In the priority-list (\code{PL}) representation, the job order
  for each machine is given. There is one line per machine and for each
  machine the orders of the jobs.
}
\seealso{
jssp.evaluate.gantt
}
