% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_gantt.R
\name{jssp.solution.start.to.gantt}
\alias{jssp.solution.start.to.gantt}
\title{Transform a Solution Represented in the van-Hoorn Solution-Start
  Method to a Gantt Chart}
\usage{
jssp.solution.start.to.gantt(data.solution.start, inst.id,
  min.job.id = 1L, get.inst.data = jssp.get.instance.data)
}
\arguments{
\item{data.solution.start}{the solution-start representation}

\item{inst.id}{the instance id}

\item{min.job.id}{the integer minimum job id to be used in the output (for
the input, it is automatically detected). By default, this is \code{1L},
but sometimes you may want to use \code{0L}.}

\item{get.inst.data}{a function obtaining the instance data for a given
instance id, by default \link{jssp.get.instance.data}}
}
\value{
the Gantt chart
}
\description{
The solution start file gives a matrix with row per per job and
  a column per machine, the values in this matrix give the start times for
  the operations. If the second row starts with a \code{10}, this indicates
  that the operation of job \code{2} on machine \code{1} starts at time
  \code{10}.
}
